Added the destination to the ufile variable
This fixes the issue where the call to client.getFile does not take into
account the destination when trying to compare the file being uploaded
to what already exists on the server. This fixes issue #19 on Github and
also solves the problem I was having where I wanted to add the files to
a subdirectory after having already added them to the root.